Update for Firefox
Updated APIs. Old API is here, with link to new the new API - https://developer.chrome.com/extensions/extension#event-onRequest

Here is the new API for Firefox, should be the same as the one for Chrome: https://developer.mozilla.org/en-US/Add-ons/WebExtensions/API/runtime/onMessage

Again not sure if I updated it right. The JSON.parse in this file is what is currently giving me all the trouble, possibly because the new API doesn't like how it's being done.

Also I removed everything related to updating the extension because Firefox was complaining about the versions not matching. I will need to look into that later.