Update changelog to indicate that this magically fixed a Firefox issue
It was ultimately the same issue: trying to POST to a non-https endpoint and getting a 301, but I don't know why Firefox was even trying to hit the http endpoint.  There's no clear sign in the code of where it would have gotten the incorrect address for a new account.